+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix `ionoscloud_certificate` data source
 - Fix `DBaaS` tests, change location for clusters creation, mark `connection_pooler` as computed
 - `certificate_id` should not be required for API Gateway resource, `custom_domains` field.
+- `cdn distribution` add metadata ipv4, ipv6 and resource_urn to resource and data source
 
 ## 6.5.5
 ### Fixes